requirements.txt
beautifulsoup4==4.5.0
bottom==1.0.2
click==6.6
dataset==0.6.4
pluginbase==0.4
python-dateutil==2.5.3
python-slugify==1.2.0
PyYAML==3.11
requests==2.10.0
youtube-dl==2016.7.24
